Review and maintenance before public release 6
Tags: [maintenance][refactor]
Compile: no tested
Tests: no tested

+++ src/noa/cpu/math/Arithmetics.h
@@ -6,7 +6,6 @@
 #pragma once
 
 #include "noa/common/Definitions.h"
-#include "noa/common/Math.h"
 #include "noa/common/Profiler.h"
 #include "noa/common/Types.h"
 
@@ -14,10 +13,10 @@ namespace noa::cpu::math::details {
     enum : int { ADD, SUBTRACT, MULTIPLY, DIVIDE, DIVIDE_SAFE };
 
     template<int OPERATION, typename T, typename U>
-    void applyValue(const T* arrays, const U* values, T* outputs, size_t elements, uint batches);
+    void applyValue(const T* arrays, const U* values, T* outputs, size_t elements, size_t batches);
 
     template<int OPERATION, typename T, typename U>
-    void applyArray(const T* arrays, const U* weights, T* outputs, size_t elements, uint batches);
+    void applyArray(const T* arrays, const U* weights, T* outputs, size_t elements, size_t batches);
 }
 
 namespace noa::cpu::math {
@@ -48,7 +47,7 @@ namespace noa::cpu::math {
     ///
     /// \note The divide*, add* and subtract* functions operate similarly, but uses the operators, /, + and -, respectively.
     template<typename T, typename U>
-    NOA_IH void multiplyByValue(const T* inputs, const U* values, T* outputs, size_t elements, uint batches) {
+    NOA_IH void multiplyByValue(const T* inputs, const U* values, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyValue<details::MULTIPLY>(inputs, values, outputs, elements, batches);
     }
@@ -65,7 +64,7 @@ namespace noa::cpu::math {
     ///
     /// \note The divide*, add* and subtract* functions operate similarly, but uses the operators, /, + and -, respectively.
     template<typename T, typename U>
-    NOA_IH void multiplyByArray(const T* inputs, const U* array, T* outputs, size_t elements, uint batches) {
+    NOA_IH void multiplyByArray(const T* inputs, const U* array, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyArray<details::MULTIPLY>(inputs, array, outputs, elements, batches);
     }
@@ -83,7 +82,7 @@ namespace noa::cpu::math {
     /// For each batch, divides \p input by a single value.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByValue().
     template<typename T, typename U>
-    NOA_IH void divideByValue(const T* inputs, U* values, T* outputs, size_t elements, uint batches) {
+    NOA_IH void divideByValue(const T* inputs, U* values, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyValue<details::DIVIDE>(inputs, values, outputs, elements, batches);
     }
@@ -91,7 +90,7 @@ namespace noa::cpu::math {
     /// For each batch, computes the element-wise division between \p input and \p array.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByArray().
     template<typename T, typename U>
-    NOA_IH void divideByArray(const T* inputs, U* array, T* outputs, size_t elements, uint batches) {
+    NOA_IH void divideByArray(const T* inputs, U* array, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyArray<details::DIVIDE>(inputs, array, outputs, elements, batches);
     }
@@ -100,7 +99,7 @@ namespace noa::cpu::math {
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByArray(),
     ///      with the additional restriction that \p U cannot be complex (cfloat_t or cdouble_t).
     template<typename T, typename U, typename = std::enable_if_t<!noa::traits::is_complex_v<U>>>
-    NOA_IH void divideSafeByArray(const T* inputs, U* array, T* outputs, size_t elements, uint batches) {
+    NOA_IH void divideSafeByArray(const T* inputs, U* array, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyArray<details::DIVIDE_SAFE>(inputs, array, outputs, elements, batches);
     }
@@ -118,7 +117,7 @@ namespace noa::cpu::math {
     /// For each batch, adds a single value to \p input.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByValue().
     template<typename T, typename U>
-    NOA_IH void addValue(const T* inputs, U* values, T* outputs, size_t elements, uint batches) {
+    NOA_IH void addValue(const T* inputs, U* values, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyValue<details::ADD>(inputs, values, outputs, elements, batches);
     }
@@ -126,7 +125,7 @@ namespace noa::cpu::math {
     /// For each batch, adds \p array to \p inputs.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByArray().
     template<typename T, typename U>
-    NOA_IH void addArray(const T* inputs, U* array, T* outputs, size_t elements, uint batches) {
+    NOA_IH void addArray(const T* inputs, U* array, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyArray<details::ADD>(inputs, array, outputs, elements, batches);
     }
@@ -144,7 +143,7 @@ namespace noa::cpu::math {
     /// For each batch, subtracts a single value to \p input.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByValue().
     template<typename T, typename U>
-    NOA_IH void subtractValue(const T* inputs, U* values, T* outputs, size_t elements, uint batches) {
+    NOA_IH void subtractValue(const T* inputs, U* values, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyValue<details::SUBTRACT>(inputs, values, outputs, elements, batches);
     }
@@ -152,7 +151,7 @@ namespace noa::cpu::math {
     /// For each batch, subtracts \p array to \p input.
     /// \see This function has the same features and restrictions than noa::cpu::math::multiplyByArray().
     template<typename T, typename U>
-    NOA_IH void subtractArray(const T* inputs, U* array, T* outputs, size_t elements, uint batches) {
+    NOA_IH void subtractArray(const T* inputs, U* array, T* outputs, size_t elements, size_t batches) {
         NOA_PROFILE_FUNCTION();
         details::applyArray<details::SUBTRACT>(inputs, array, outputs, elements, batches);
     }
